Don't read structs directly
I'm now on a mission to avoid undefined and implementation-defined
behavior in libbmpread.  This first part of that journey is the biggest:
not reading structs directly out of the file, but correctly parsing them
as little-endian data.  Not only does this make libbmpread theoretically
more portable, but it also obviates the need for byte swapping code
entirely.