Improve ReadPalette implementation
I haven't measured performance, but this is a cleaner implementation.
Instead of reading each byte directly into the palette_entry struct, we
read all 4 components of each color in one fread, then copy bytes into
the struct.  It feels better this way, and who knows, it might make it
easier for the compiler to optimize.