Random smattering of changes to reduce U.B.
This cleans up a little type handling etc. with the goal of making it
easier to verify that no undefined behavior lurks there.  It also adds a
couple checks for U.B. conditions that we can catch ahead of time.

There's still work to do--for example we still use int in the absence of
an uber-portable ssize_t type (not in the ANSI C spec, just POSIX).