I added alpha loading and some examples.
The alpha channel will be loaded if the flag BMPREAD_LOAD_ALPHA in enabled. In case of a bitmap that doesn't provide a alpha channel all alpha values will be set to 0xff.

The examples added are "art" and can be used by everyone for everything... Except selling them under a different name.(No one would buy them!!)